+++ driver/sql/postgres/pgjournal/doc.go
@@ -1,0 +1,3 @@
+// Package pgjournal provides an implementation of [journal.Store] that persists
+// records in a PostgreSQL database.
+package pgjournal