+++ driver/sql/postgres/pgkv/doc.go
@@ -1,0 +1,3 @@
+// Package pgkv provides an implementation of [kv.Store] that persists data in a
+// PostgreSQL database.
+package pgkv