+++ driver/sql/postgres/pgkv/keyspace.go
@@ -1,4 +1,4 @@
-package postgres
+package pgkv
 
 import (
 	"context"
@@ -6,21 +6,6 @@ import (
 
 	"github.com/dogmatiq/persistencekit/kv"
 )
-
-// KeyValueStore is an implementation of [kv.Store] that stores keyspaces in a
-// PostgreSQL database.
-type KeyValueStore struct {
-	DB *sql.DB
-}
-
-// Open returns the keyspace with the given name.
-func (s *KeyValueStore) Open(_ context.Context, name string) (kv.Keyspace, error) {
-	// TODO: consider creating a separate table partition for each keyspace
-	return &keyspace{
-		Name: name,
-		DB:   s.DB,
-	}, nil
-}
 
 type keyspace struct {
 	Name string