+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased
+
+### Changed
+
+- **[BC]** Moved `postgres.JournalStore` to `pgjournal.Store`
+- **[BC]** Moved `postgres.KeyValueStore` to `pgkv.Store`
+
 ## [0.1.0] - 2023-10-16
 
 ### Added